Implement JCR-2454 registerNodeTypes via DAVEX
This implements registration of node types via DAVEX remoting.
The client creates a CND version of the node type definition and
passes this on to the server that parses the cnd and registers
the node type

Also includes the allowUpdate flag in the request.

PROPFIND for nodetypes-cnd was implemented. In order to use
CompactNodeTypeDefWriter jackrabbit-jcr-server needs to depend on
jackrabbit-spi and jackrabbit-spi-commons (not sure if that's a
problem).

Integration tests for registering node types enabled and running